requirements.txt
streamlit==1.28.1
pandas==2.0.3
folium==0.14.0
streamlit-folium==0.15.0
simplekml==1.3.6
Pillow==10.0.0